Integração Gerencia Net

+++ database/migrations/2018_10_25_030418_create_planos_table.php
@@ -16,7 +16,7 @@ class CreatePlanosTable extends Migration
             $table->integer('quant_anuncios');
             $table->integer('super_destaques');
             $table->integer('destaques');
-            $table->decimal('valor_mensal', 10,2);
+            $table->integer('valor_mensal');
             $table->boolean('captacao')->default(false);            
             $table->timestamps();
         });
